+++ cloud/disk_manager/pkg/admin/ya.make
@@ -5,7 +5,6 @@ SRCS(
     disks.go
     filesystem.go
     images.go
-    nbs.go
     operations.go
     placement_group.go
     pools.go
